Add DVD file stat from version 0.9.7 to libdvdread
A new function to do an informational read of the DVD file handle was
introduced to libdvdread 0.9.7 (or was it .6). Since libdvdread at
mplayerhq imported an older version of dvdread/dvdnav this part of
the API was not available. This commit places that function into
the library making this library complete in reference to 0.9.7.

git-svn-id: svn://svn.mplayerhq.hu/dvdnav/trunk/libdvdread@1178 625a9de0-492d-0410-920c-a7e632af1907