Provide BUP file support for more issues.
The BUP file is only opened when the IFO file open fails. We have a
few times where file corruption could happen and we could use the
BUP instead. This patch attempts to address this by trying to
open the BUP if there is any reported error w/ the IFO. Inspiration
for this patch came from Rich E, thanks for the detailed bug
report and attempts at using earlier patches.

git-svn-id: svn://svn.mplayerhq.hu/dvdnav/trunk/libdvdread@1215 625a9de0-492d-0410-920c-a7e632af1907