Inhibit double free of parental management information
If reading the country data fails the point to the parental
management information is not nulled out. So it is possible for
the same structure to be freed twice. To inhibit this whenever
the country data read fails the structure pointer is assigned
null. Patch submitted by Frederic Marchal, fmarchal at
perso dot be. Thanks for being patient and reworking the patch
when requested.

git-svn-id: svn://svn.mplayerhq.hu/dvdnav/trunk/libdvdread@1220 625a9de0-492d-0410-920c-a7e632af1907