Detect physical drive or image on Windows w/ mingw
It's purpose is to detect if an input path is a physical drive (on
windows), and if so attempt to open it as an image.  The current code
will only attempt to open the path as an image if the path contains no
directory components (e.g. "X:" or "X:\").  So if the path instead is
"X:\VIDEO_TS", it will fail to open it as an image and instead open it
as a directory tree and css decryption will not work. This makes the
behavior more similar to what happens on other platforms.